REST_API/requirements.txt
django==3.1.7 
djangorestframework==3.12.4
django-cors-headers==3.5.0


